feat(extension-chrome): refactor unit test
