+++ packages/extension-chrome/src/rpc/schema/__mocks__/index.ts
@@ -1,0 +1,3 @@
+export function bindSchemaValidator(method: unknown, handler: unknown): unknown {
+  return handler;
+}
